Re-add Nub.cs and move logic from NormalNub.cs to it.
